popVideo

+++ backend/src/main/java/com/example/hdt/ServiceImpl/PatientImpl.java
@@ -100,7 +100,7 @@ public class PatientImpl{
         mongoTemplate.updateFirst(query,update,Patient.class);
     }
 
-    public void updateTotalHour(String patientId, Double totalHour, Double weekHour){
+    public void updateTotalHour(String patientId, int totalHour, int weekHour){
         Query query = new Query(Criteria.where("PatientID").is(patientId));
         Update update = new Update().set("TotalExerciseHours",totalHour).set("WeekExerciseHours",weekHour);
         mongoTemplate.updateFirst(query,update,Patient.class);
